don't mark `declval` as `{.compileTime.}`
`declval` is meant to be used to determine types of expressions, and never be executed (https://github.com/status-im/nim-stew/pull/33). However it's marked `{.compileTime.}`, which would normally make it always execute during constant folding, but due to a bug in Nim (https://github.com/nim-lang/Nim/issues/10753), generic procs like `declval` do not undergo constant folding. So the `{.compileTime.}` annotation is removed for this to work when the bug in Nim is fixed.